dr, bdr is fixed, drawing works better, msg() and debug() added
A bug, where the DR and BDR would be the same color is fixed now.
msg() and debug() functions have been implemented and `-d` a `-s`
command line flags are not supported for debuging and silencing warnings
or error output.

There still needs to be more work done on the drawing part, mainly on
drawing Multi Access areas. Currently, only straight lines from router
to router are drawn. Paths through the switch are ignored and they need
to be shown.